backend/requirements.txt
Django==2.2.6
pytz==2020.1
sqlparse==0.3.1
psycopg2-binary==2.8.3
gunicorn==19.9.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
django-phonenumber-field==4.0.0
phonenumbers==8.12.7
django-address==0.2.4
django-cors-headers==3.4.0